+++ Sources/Internal/Render/RHI/Common/MCPP/directive.cpp
@@ -1348,7 +1348,7 @@ DEFBUF **   look_prev(
 
     for (hash = 0, np = name; *np != EOS; )
         hash += *np++;
-    hash += s_name = (size_t)(np - name);
+    hash += (int)(s_name = (size_t)(np - name));
     s_name++;
     prevp = & symtab[ hash & SBMASK];
     *cmp = -1;                              /* Initialize           */
