+++ Sources/Internal/Render/RHI/Common/MCPP/system.cpp
@@ -2486,7 +2486,7 @@ static char *   norm_path(
     inf = inf && (mcpp_debug & PATH);       /* Output information   */
 
     strcpy( slbuf1, dir);                   /* Include directory    */
-    len = strlen( slbuf1);
+    len = static_cast<int>(strlen( slbuf1));
     if (fname && len && slbuf1[ len - 1] != PATH_DELIM) {
         slbuf1[ len] = PATH_DELIM;          /* Append PATH_DELIM    */
         slbuf1[ ++len] = EOS;
@@ -2548,7 +2548,7 @@ static char *   norm_path(
                     , dir, fname ? fname : null, slbuf1);
     }
 #endif
-    len = strlen( slbuf1);
+    len = static_cast<int>(strlen( slbuf1));
     start = norm_name = xmalloc( len + 1);  /* Need a new buffer    */
     strcpy( norm_name, slbuf1);
 #if SYS_FAMILY == SYS_WIN
